Move eh_unwind_resume into CrateContext
Also improves cache quality.
